requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Validation and serialization
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2

# CORS support
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# Monitoring and logging
psutil==5.9.6